python_app/requirements.txt
PySide6>=6.6
Pillow>=10.2
opencv-python>=4.9
numpy>=1.26
python-piexif>=1.1.3
